+++ src/main/java/com/example/avenueMailExample/service/UserService.java
@@ -1,6 +1,5 @@
 package com.example.avenueMailExample.service;
 
-
 import com.example.avenueMailExample.entity.User;
 import com.example.avenueMailExample.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,13 +15,13 @@ public class UserService {
     public UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
-    public User findById(int id){
+    public User findById(Long id){
         return userRepository.getOne(id);
     }
     public User saveUser(User user){
         return userRepository.save(user);
     }
-    public void deleteUser(int id){
+    public void deleteUser(Long id){
         userRepository.deleteById(id);
     }
     public List<User> findAll(){
